+++ IocHelpers/TinyIoc/TinyIocHandlerFactory.cs
@@ -1,0 +1,71 @@
+#region Licence
+/* The MIT License (MIT)
+Copyright © 2015 Ian Cooper <ian_hammond_cooper@yahoo.co.uk>
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the “Software”), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE. */
+
+#endregion
+
+using System;
+using System.Collections.Generic;
+using paramore.brighter.commandprocessor;
+using TinyIoC;
+
+namespace paramore.contrib.brighter.iocHelpers.TinyIoc
+{
+    internal class TinyIocHandlerFactory : IAmAHandlerFactory, IAmASubscriberRegistry
+    {
+        private readonly TinyIoCContainer _container;
+        private readonly SubscriberRegistry _registry;
+
+        public TinyIocHandlerFactory(TinyIoCContainer container)
+        {
+            _container = container;
+            _registry = new SubscriberRegistry();
+        }
+
+        public IHandleRequests Create(Type handlerType)
+        {
+            return (IHandleRequests)_container.Resolve(handlerType);
+        }
+
+        public void Release(IHandleRequests handler)
+        {
+            var disposable = handler as IDisposable;
+
+            if (disposable != null)
+            {
+                disposable.Dispose();
+            }
+        }
+
+        public IEnumerable<Type> Get<T>() where T : class, IRequest
+        {
+            return _registry.Get<T>();
+        }
+
+        public void Register<TRequest, TImplementation>()
+            where TRequest : class, IRequest
+            where TImplementation : class, IHandleRequests<TRequest>
+        {
+            _container.Register<TImplementation>().AsMultiInstance();
+            _registry.Register<TRequest, TImplementation>();
+        }
+    }
+}
